+++ app/build.gradle
@@ -26,7 +26,7 @@ android {
     }
 }
 
-repositories{
+repositories {
     mavenCentral()
 }
 
@@ -42,5 +42,6 @@ dependencies {
     compile 'com.squareup.okhttp3:okhttp:3.6.0'
     compile 'com.android.support:cardview-v7:25.3.1'
     compile 'com.android.support.constraint:constraint-layout:1.0.2'
+    compile 'com.android.support:support-v4:25.3.1'
     testCompile 'junit:junit:4.12'
 }
