+++ app/src/main/java/com/arturgiro/android/popularmovies/activities/MainActivity.java
@@ -1,50 +1,24 @@
 package com.arturgiro.android.popularmovies.activities;
 
-import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.database.Cursor;
-import android.graphics.Rect;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.annotation.DimenRes;
-import android.support.annotation.NonNull;
-import android.support.design.widget.CoordinatorLayout;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.ProgressBar;
-import android.widget.TextView;
 
 import com.arturgiro.android.popularmovies.R;
 import com.arturgiro.android.popularmovies.data.Movie;
-import com.arturgiro.android.popularmovies.data.MoviesContract;
-import com.arturgiro.android.popularmovies.network.AsyncTaskDelegate;
-import com.arturgiro.android.popularmovies.network.FetchMoviesTask;
-import com.arturgiro.android.popularmovies.network.NetworkUtils;
-import com.arturgiro.android.popularmovies.views.EndlessRecyclerViewScrollListener;
-import com.arturgiro.android.popularmovies.views.adapters.MoviesAdapter;
-
-import java.util.ArrayList;
 
 import static com.arturgiro.android.popularmovies.data.Movie.MOVIE_IDENTIFIER;
 
-public class MainActivity extends AppCompatActivity implements MoviesAdapter.MoviesAdapterOnClickHandler, SharedPreferences.OnSharedPreferenceChangeListener, AsyncTaskDelegate{
+public class MainActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener, GridFragment.OnGridFragmentInteractionListener {
+
+    private boolean mTwoPane;
 
     private static final String TAG = MainActivity.class.getSimpleName();
-
-    private RecyclerView mRecyclerView;
-    private TextView mErrorMessageDisplay;
-    private ProgressBar mLoadingIndicator;
-    private CoordinatorLayout coordinatorLayout;
-
-    private MoviesAdapter mMoviestAdapter;
-
-    private EndlessRecyclerViewScrollListener mScrollListener;
+    private static final String DETAILFRAGMENT_TAG = "DFTAG";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -54,83 +28,28 @@ public class MainActivity extends AppCompatActivity implements MoviesAdapter.Mov
         PreferenceManager.setDefaultValues(this, R.xml.preferences, false);
         SharedPreferences sharedPref = PreferenceManager.getDefaultSharedPreferences(this);
         sharedPref.registerOnSharedPreferenceChangeListener(this);
-
-        mRecyclerView = (RecyclerView) findViewById(R.id.recyclerview_movies);
-        mErrorMessageDisplay = (TextView) findViewById(R.id.tv_error_message_display);
-        mLoadingIndicator = (ProgressBar) findViewById(R.id.pb_loading_indicator);
-        coordinatorLayout = (CoordinatorLayout) findViewById(R.id.coordinatorLayout);
-
-
-        GridLayoutManager layoutManager = new GridLayoutManager(this, 2);//TODO VErificar esse número
-        mRecyclerView.setLayoutManager(layoutManager);
-        ItemOffsetDecoration itemDecoration = new ItemOffsetDecoration(this, R.dimen.item_offset);
-        mRecyclerView.addItemDecoration(itemDecoration);
-
-        mRecyclerView.setHasFixedSize(true);
-
-        mMoviestAdapter = new MoviesAdapter(this);
-        mRecyclerView.setAdapter(mMoviestAdapter);
-
-        // Retain an instance so that you can call `resetState()` for fresh searches
-       mScrollListener = new EndlessRecyclerViewScrollListener(layoutManager) {
-            @Override
-            public void onLoadMore(int page, int totalItemsCount, RecyclerView view) {
-                // Triggered only when new data needs to be appended to the list
-                // Add whatever code is needed to append new items to the bottom of the list
-                loadMoviesData(page);
-            }
-        };
-
-        // Adds the scroll listener to RecyclerView
-        mRecyclerView.addOnScrollListener(mScrollListener);
-
-        loadMoviesData(1);
-    }
-
-    /**
-     * This method will get the user's preferred sort criteria, and then tell some
-     * background method to get the movies data in the background.
-     */
-    private void loadMoviesData(final int page) {
-
-        showMoviesDataView();
-
-        SharedPreferences sharedPref = PreferenceManager.getDefaultSharedPreferences(this);
         String sortMethod = sharedPref.getString(getString(R.string.pref_order_key), "");
-        if(sortMethod.compareTo(getString(R.string.pref_sort_label_favorite)) == 0){
-            if(page == 1)//não usa paginação para os filmes salvos no banco
-                refreshMovies(getFavorites());
+
+        if (findViewById(R.id.detail_container) != null) {
+            // The detail container view will be present only in the large-screen layouts
+            // (res/layout-sw600dp). If this view is present, then the activity should be
+            // in two-pane mode.
+            mTwoPane = true;
+            // In two-pane mode, show the detail view in this activity by
+            // adding or replacing the detail fragment using a
+            // fragment transaction.
+            if (savedInstanceState == null) {
+                getSupportFragmentManager().beginTransaction()
+                        .replace(R.id.detail_container, new DetailFragment(), DETAILFRAGMENT_TAG)
+                        .commit();
+            }
+        } else {
+            mTwoPane = false;
         }
-        else
-            if(NetworkUtils.isNetworkConnected(this)) {
-                new FetchMoviesTask(this).execute(sortMethod, String.valueOf(page));
-            }
-            else {
-                Snackbar snackbar = Snackbar.make(coordinatorLayout, getString(R.string.no_internet_connection), Snackbar.LENGTH_INDEFINITE);
 
-                snackbar.setAction(getString(R.string.retry), new View.OnClickListener() {
-                    //Ao clicar na snackbar, uma nova tentativa de atualizar a lista é efetuada :-)
-                    @Override
-                    public void onClick(View view) {
-                        loadMoviesData(page);
-                    }
-                });
-                snackbar.show();
-            }
-    }
-
-    private void showMoviesDataView() {
-       //Hide the error message
-        mErrorMessageDisplay.setVisibility(View.INVISIBLE);
-        //show the movies
-        mRecyclerView.setVisibility(View.VISIBLE);
-    }
-
-    private void showErrorMessage() {
-        /* First, hide the currently visible data */
-        mRecyclerView.setVisibility(View.INVISIBLE);
-        /* Then, show the error */
-        mErrorMessageDisplay.setVisibility(View.VISIBLE);
+        GridFragment fragment =  ((GridFragment)getSupportFragmentManager()
+                .findFragmentById(R.id.fragment_grid));
+        fragment.setSortMethod(sortMethod);
     }
 
     @Override
@@ -153,98 +72,44 @@ public class MainActivity extends AppCompatActivity implements MoviesAdapter.Mov
         return super.onOptionsItemSelected(item);
     }
 
-    /**
-     * This method is overridden by our MainActivity class in order to handle RecyclerView item
-     * clicks.
-     *
-     * @param movie The movie that was clicked
-     */
-    @Override
-    public void onClick(Movie movie) {
-        Context context = this;
-        Class destinationClass = MovieDetailActivity.class;
-        Intent intentToStartDetailActivity = new Intent(context, destinationClass);
-        intentToStartDetailActivity.putExtra(MOVIE_IDENTIFIER, movie);
-        startActivity(intentToStartDetailActivity);
-    }
-
     @Override
     public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
 
-        if (key.equals(getString(R.string.pref_order_key))) {
-            // clear the array of data
-            mMoviestAdapter.resetMovieData();
-            // reset endless scroll listener when performing a new search
-            mScrollListener.resetState();
-
-            loadMoviesData(1);
-
-        }
+//        if (key.equals(getString(R.string.pref_order_key))) {
+//            // clear the array of data
+//            mMoviestAdapter.resetMovieData();
+//            // reset endless scroll listener when performing a new search
+//            mScrollListener.resetState();
+//
+//            loadMoviesData(1);
+//
+//        }
     }
 
-    private ArrayList<Movie> getFavorites() {
-        Cursor cursor = getContentResolver().query(MoviesContract.MovieEntry.CONTENT_URI, null, null, null, null);
-        ArrayList<Movie> ret = new ArrayList<Movie>();
-        try {
-            while (cursor.moveToNext()) {
-                Movie m = new Movie(cursor.getInt(cursor.getColumnIndex(MoviesContract.MovieEntry._ID)),
-                                    cursor.getString(cursor.getColumnIndex(MoviesContract.MovieEntry.COLUMN_TITLE)),
-                                    cursor.getString(cursor.getColumnIndex(MoviesContract.MovieEntry.COLUMN_POSTER_PATH)),
-                                    cursor.getString(cursor.getColumnIndex(MoviesContract.MovieEntry.COLUMN_OVERVIEW)),
-                                    cursor.getString(cursor.getColumnIndex(MoviesContract.MovieEntry.COLUMN_USER_RATING)),
-                                    cursor.getString(cursor.getColumnIndex(MoviesContract.MovieEntry.COLUMN_RELEASE_DATE))
-                                    );
-                ret.add(m);
-            }
-        } finally {
-            cursor.close();
-        }
-        return ret;
-    }
-
+    //callback method from the grid fragment
     @Override
-    public void processStart() {
-        mLoadingIndicator.setVisibility(View.VISIBLE);
-    }
+    public void onItemSelected(Movie movie) {
+        if (mTwoPane) {
+            // In two-pane mode, show the detail view in this activity by
+            // adding or replacing the detail fragment using a
+            // fragment transaction.
+            Bundle args = new Bundle();
+            args.putParcelable(MOVIE_IDENTIFIER, movie);
 
-    @Override
-    public void processFinish(Object output) {
-        mLoadingIndicator.setVisibility(View.INVISIBLE);
+            DetailFragment fragment = new DetailFragment();
+            fragment.setArguments(args);
 
-        ArrayList<Movie> movieData = (ArrayList<Movie>)output;
-        refreshMovies(movieData);
-    }
-
-    private void refreshMovies(ArrayList<Movie> movieData) {
-
-        if (movieData != null) {
-            showMoviesDataView();
-            mMoviestAdapter.setMoviesData(movieData);
+            getSupportFragmentManager().beginTransaction()
+                    .replace(R.id.detail_container, fragment, DETAILFRAGMENT_TAG)
+                    .commit();
         } else {
-            showErrorMessage();
-        }
-
-    }
-
-    public class ItemOffsetDecoration extends RecyclerView.ItemDecoration {
-
-        private int mItemOffset;
-
-        public ItemOffsetDecoration(int itemOffset) {
-            mItemOffset = itemOffset;
-        }
-
-        public ItemOffsetDecoration(@NonNull Context context, @DimenRes int itemOffsetId) {
-            this(context.getResources().getDimensionPixelSize(itemOffsetId));
-        }
-
-        @Override
-        public void getItemOffsets(Rect outRect, View view, RecyclerView parent,
-                                   RecyclerView.State state) {
-            super.getItemOffsets(outRect, view, parent, state);
-            outRect.set(mItemOffset, mItemOffset, mItemOffset, mItemOffset);
+            Class destinationClass = DetailActivity.class;
+            Intent intentToStartDetailActivity = new Intent(this, destinationClass);
+            intentToStartDetailActivity.putExtra(MOVIE_IDENTIFIER, movie);
+            startActivity(intentToStartDetailActivity);
         }
     }
+
 //TODO Adaptar layout para tablet
 // TODO recreating activities using onSaveInstanceState.  If your app crashes when the screen is rotated, you will not pass the project.
