+++ app/src/main/java/com/arturgiro/android/popularmovies/network/LoaderReviewDelegate.java
@@ -1,0 +1,6 @@
+package com.arturgiro.android.popularmovies.network;
+
+public interface LoaderReviewDelegate {
+    void loadReviewStart();
+    void loadReviewFinish(Object output);
+}
