+++ app/src/main/java/com/arturgiro/android/popularmovies/network/LoaderVideoDelegate.java
@@ -1,0 +1,6 @@
+package com.arturgiro.android.popularmovies.network;
+
+public interface LoaderVideoDelegate {
+    void loadVideoStart();
+    void loadVideoFinish(Object output);
+}
