+++ app/src/main/res/layout/fragment_detail.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/coordinator_layout"
+    android:id="@+id/relative_layout_detail"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true">
@@ -18,7 +18,7 @@
             android:layout_height="match_parent"
             android:layout_margin="8dp"
             android:orientation="vertical"
-            tools:context="com.arturgiro.android.popularmovies.activities.MovieDetailActivity">
+            tools:context="com.arturgiro.android.popularmovies.activities.DetailActivity">
 
             <LinearLayout
                 android:layout_width="match_parent"
