+++ app/src/main/res/values/strings.xml
@@ -91,4 +91,7 @@
 
     <string name="add_favorite">Added to favorites</string>
     <string name="remove_favorite">Removed from favorites</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
